requirements.txt
# Core Python packages for data analysis and visualization
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0

# HTTP and async requests
requests>=2.25.0
httpx>=0.24.0
aiohttp>=3.8.0

# Progress bars and CLI
tqdm>=4.62.0

# File path handling
pathlib>=1.0.1

# AI/ML and API clients
openai>=1.0.0

# Environment and configuration
python-dotenv>=0.19.0

# Network and IP handling
ipwhois>=1.2.0
ping3>=4.0.0

# Database connectivity
psycopg2-binary>=2.9.0

# Google Cloud services
google-cloud-bigquery>=3.0.0

# Web scraping and domain parsing
tldextract>=3.4.0

# SVG drawing for visualizations
drawsvg>=2.0.0